% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg.R
\name{angle_to_net}
\alias{angle_to_net}
\title{Compute Angle to Goal}
\usage{
angle_to_net(x, y)
}
\arguments{
\item{x}{A numeric vector of the x coordinates}

\item{y}{A numeric vector of the y coordinates}
}
\value{
A numeric vector of the computed angles
}
\description{
Compute the angles from points to the centre of the net
}
