% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_values.R
\name{compute_ripp}
\alias{compute_ripp}
\title{Compute RIPP}
\usage{
compute_ripp(data)
}
\arguments{
\item{data}{A data frame of event data}
}
\value{
A numeric vector of RIPP values
}
\description{
Compute RIPP given event data frame
}
