% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg.R
\name{distance_to_net}
\alias{distance_to_net}
\title{Compute Distance to Goal}
\usage{
distance_to_net(x, y)
}
\arguments{
\item{x}{A numeric vector of the x coordinates}

\item{y}{A numeric vector of the y coordinates}
}
\value{
A numeric vector of the computed distances
}
\description{
Compute the distances from points to the centre of the net
}
