% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{event_data}
\alias{event_data}
\title{Event Data}
\format{
An object of class \code{data.frame} with 61493 rows and 23 columns.
}
\source{
\url{https://github.com/bigdatacup/Big-Data-Cup-2021}
}
\usage{
event_data
}
\description{
A dataset containing the merged records from the three event data sources
provided in the 2022 Big Data Cup.
}
\keyword{datasets}
