% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{player_ripp_vs_goals}
\alias{player_ripp_vs_goals}
\title{Plot Players' RIPP per Game vs Goals per Game}
\usage{
player_ripp_vs_goals(event_values)
}
\arguments{
\item{event_values}{A data frame of event data}
}
\value{
A formatted ggplot
}
\description{
Build plot of players' average RIPP per game vs their goals per game
}
