% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{player_vs_average}
\alias{player_vs_average}
\title{Plot Players' RIPP per Game vs Goals per Game}
\usage{
player_vs_average(event_values, our_player)
}
\arguments{
\item{event_values}{A data frame of event data}

\item{our_player}{The name of the player of interest}
}
\value{
A formatted ggplot
}
\description{
Build bar plot of player's RIPP per game per event type vs the average player
}
