% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg.R
\name{predict_xg}
\alias{predict_xg}
\title{Predict Expected Goals}
\usage{
predict_xg(x, y)
}
\arguments{
\item{x}{A numeric vector of the x coordinates}

\item{y}{A numeric vector of the y coordinates}
}
\value{
A numeric vector of the computed expected goals values
}
\description{
Predict the expected goals for shots with start coordinates (x, y)
}
