% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg.R
\name{prepare_xg_features}
\alias{prepare_xg_features}
\title{Prepare Expected Goals Features}
\usage{
prepare_xg_features(data)
}
\arguments{
\item{data}{A data frame containing \code{X.Coordinate} and \code{Y.Coordinate} columns}
}
\value{
The same data frame with additional columns for any expected goals features
}
\description{
Add any features needed for the expected goals model
}
